debug websocker for prod
