+++ docker-compose.yml
@@ -1,26 +1,21 @@
 services:
   web:
-    # build: .
-    image: atemndobs/viewer-pwa:v0.3
+    build: .
+    # image: atemndobs/viewer-pwa-amd64:v1.1
     container_name: viewer-pwa
     ports:
       - "3400:3000"
-    # volumes:
-      # - .:/app
     environment:
       - NODE_ENV=production
       - YOUTUBE_API_KEY=AIzaSyC0kQQzDSk_gEpjmcHfHlTdwmnQ8kKGQIo
+      - WEBSOCKET_URL=wss://viewer.atemkeng.de/ws
 
-
-
-# services:
-#   web:
-#     build: .
-#     container_name: youtube-viewer
-#     ports:
-#       - "3400:3000"
-#       - "3001:3001"
-#     volumes:
-#       - .:/app
-#     environment:
-#       - NODE_ENV=production
+  websocket-server:
+    build: ./server
+    # image: atemndobs/viewer-ws-amd64:v0.2
+    container_name: websocket-server
+    ports:
+      - "8681:8681"
+    environment:
+      - NODE_ENV=production
+      - WEBSOCKET_URL=wss://viewer.atemkeng.de/ws
