+++ nginx/nginx.conf
@@ -1,0 +1,54 @@
+worker_processes 1;
+
+events {
+  worker_connections 1024;
+}
+
+http {
+  include       mime.types;
+  default_type  application/octet-stream;
+  
+  # Log format for easier debugging
+  log_format main '$remote_addr - $remote_user [$time_local] "$request" '
+                  '$status $body_bytes_sent "$http_referer" '
+                  '"$http_user_agent" "$http_x_forwarded_for"';
+
+  access_log /var/log/nginx/access.log main;
+  error_log /var/log/nginx/error.log warn;
+
+  # HTTP to HTTPS redirect
+  server {
+    listen 80;
+    server_name viewer.atemkeng.de;
+    return 301 https://$host$request_uri;
+  }
+
+  # HTTPS server for secure WebSocket and website
+  server {
+    listen 443 ssl;
+    server_name viewer.atemkeng.de;
+
+    ssl_certificate /etc/nginx/certs/your_certificate.crt; # Update with your certificate path
+    ssl_certificate_key /etc/nginx/certs/your_private.key; # Update with your private key path
+
+    location / {
+      proxy_pass http://web:3000;
+      proxy_set_header Host $host;
+      proxy_set_header X-Real-IP $remote_addr;
+      proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+      proxy_set_header X-Forwarded-Proto $scheme;
+    }
+
+    # WebSocket reverse proxy
+    location /ws {
+      proxy_pass http://websocket-server:8681;
+      proxy_http_version 1.1;
+      proxy_set_header Upgrade $http_upgrade;
+      proxy_set_header Connection "upgrade";
+      proxy_set_header Host $host;
+      proxy_set_header X-Real-IP $remote_addr;
+      proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+      proxy_set_header X-Forwarded-Proto $scheme;
+    }
+  }
+}
