+++ server/Dockerfile
@@ -1,0 +1,20 @@
+# Use Node.js 18 as the base image
+FROM node:18
+
+# Set the working directory
+WORKDIR /app
+
+# Copy package.json and package-lock.json
+COPY package*.json ./
+
+# Install dependencies
+RUN npm install
+
+# Copy the rest of the application code
+COPY . .
+
+# Expose the WebSocket port
+EXPOSE 8681
+
+# Command to run the WebSocket server
+CMD ["node", "websocketServer.js"]
